+++ src/viem/Actions/token.test.ts
@@ -349,11 +349,12 @@ describe('getMetadata', () => {
   })
 
   test('behavior: quote token', async () => {
-    const metadata = await actions.token.getMetadata(client, {
-      token: TokenIds.defaultQuoteToken,
-    })
+    {
+      const metadata = await actions.token.getMetadata(client, {
+        token: TokenIds.defaultQuoteToken,
+      })
 
-    expect(metadata).toMatchInlineSnapshot(`
+      expect(metadata).toMatchInlineSnapshot(`
       {
         "currency": "USD",
         "decimals": 6,
@@ -362,6 +363,23 @@ describe('getMetadata', () => {
         "totalSupply": 0n,
       }
     `)
+    }
+
+    {
+      const metadata = await actions.token.getMetadata(client, {
+        token: Addresses.defaultQuoteToken,
+      })
+
+      expect(metadata).toMatchInlineSnapshot(`
+      {
+        "currency": "USD",
+        "decimals": 6,
+        "name": "linkingUSD",
+        "symbol": "linkingUSD",
+        "totalSupply": 0n,
+      }
+    `)
+    }
   })
 
   test('behavior: custom token (id)', async () => {